# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Accessories Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount: number;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  imageUrls: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  sku?: string | null;
  stock: number;
  featured: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  tags: string[];
  createdAt: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ProductStats {
  totalProducts: number;
  totalCategories: number;
  featuredCount: number;
  outOfStock: number;
  avgPrice: number;
  /** @nullable */
  topCategory?: string | null;
}

export interface CreateProductRequest {
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  imageUrls?: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  sku?: string | null;
  stock: number;
  featured?: boolean;
  tags?: string[];
}

export interface UpdateProductRequest {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  compareAtPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  imageUrls?: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  stock?: number | null;
  /** @nullable */
  featured?: boolean | null;
  tags?: string[];
}

export interface Review {
  id: number;
  productId: number;
  reviewerName: string;
  rating: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface CreateReviewRequest {
  reviewerName: string;
  rating: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  comment?: string | null;
}

export interface CartItem {
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
  subtotal: number;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartRequest {
  sessionId: string;
  productId: number;
  quantity: number;
}

export interface UpdateCartItemRequest {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  subtotal: number;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  shippingAddress: string;
  items: OrderItem[];
  total: number;
  status: string;
  createdAt: string;
}

export interface CreateOrderRequest {
  sessionId: string;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  shippingAddress: string;
}

export interface UpdateOrderStatusRequest {
  status: string;
}

export type GetProductsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  minPrice?: number | null;
  /**
   * @nullable
   */
  maxPrice?: number | null;
  /**
   * @nullable
   */
  featured?: boolean | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetCartParams = {
  sessionId: string;
};