# lib/api-zod/src/generated/types/addToCartRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Accessories Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface AddToCartRequest {
  sessionId: string;
  productId: number;
  quantity: number;
}