# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Accessories Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
  subtotal: number;
}