# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Accessories Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderRequest {
  sessionId: string;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  shippingAddress: string;
}