# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Accessories Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  shippingAddress: string;
  items: OrderItem[];
  total: number;
  status: string;
  createdAt: string;
}